backend/requirements.txt
fastapi~=0.109.0
uvicorn[standard]~=0.27.0
sqlalchemy[asyncio]~=2.0.25
aiosqlite~=0.19.0
asyncpg~=0.29.0
pydantic~=2.5.0
pydantic-settings~=2.1.0
python-jose[cryptography]~=3.3.0
passlib[bcrypt]~=1.7.4
bcrypt>=4.0.0,<4.1.0
python-dotenv~=1.0.0
httpx~=0.26.0
python-multipart~=0.0.6
email-validator~=2.1.0
apscheduler~=3.10.0
slowapi~=0.1.9
alembic~=1.13.0
Pillow~=10.0.0
bleach~=6.1.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
